bulk service added, gem refactored
