+++ lib/bing-ads-reporting.rb
@@ -1,4 +1,6 @@
-require_relative 'bing-ads-reporting/service'
+require_relative 'bing-ads-reporting/core_service'
+require_relative 'bing-ads-reporting/ad_insight_service'
+require_relative 'bing-ads-reporting/bulk_service'
 require_relative 'bing-ads-reporting/version'
 require 'logger'
